requirements.txt
environs==11.0.0
python-telegram-bot==13.7
requests==2.32.3
vk-api==11.9.9
